Introduce pchan otherend_close and remove poll queue in rx_worker
Currently in rx_worker, pchan->hyp_data is used to check the health
of the pchan itself. It is cleaned in vhost_hab_release() when closing
the vhost-dev file in vhost-user-qti, which is after reset_owner ioctl.

When handling GVM restart/crash scenario, above timing is too late to
stop pchan activities. It causes unnecessary send messages be added
into send_list in vh_pchan, then wake up vhost worker meaninglessly as
the remote side is already dead and cannot receive anything.

Thus, a new status, otherend_closed, is introduced in pchan and checked
in each pchan send. The status is set to 0 during the initialization
and set to 1 in reset_owner ioctl.

Meanwhile, the vhost_poll_queue in rx_worker is removed. Because the
vhost framework will unset the VRING_USED_F_NO_NOTIFY flag in used ring
when there is no available inbuf to use. It will make the queue kickable.

The message to send is still left in the send_list. Whenever a inbuf is
refilled, virtio HAB driver will trigger a kick which will wake up the
vhost worker via vhost_poll of the rx virtqueue. The message will be
taken care in that call. There is no need to poll the rx queue by HAB
driver itself and entering busy wait mode.

Signed-off-by: lixiang <quic_lixian@quicinc.com>
Change-Id: I85d84000528300e12a66689c246b2d483626aeb7