HAB: Reduce the log frequency during GVM restart
When a GVM restart or shutdown is triggered, the HAB driver monitors
the remaining alive virtual channels and waits for their closure.

During this waiting period, error logs will be printed periodically
to indicate which channel is still alive. However, the frequency of
the log is too high, which corrupts the Linux dmesg output.

Add a 2-second interval for the log printing.

Change-Id: I368ce0adea127199f923b6812b81d8dc20754eb4
Signed-off-by: lixiang <quic_lixian@quicinc.com>