HAB: Correct the pchan searching method in HAB open
In the current initialization phase, HAB devices are created with a
full set, including domains that are not enabled such as 701, 801.
If the physical channel is not enabled or is accidentally freed,
the pchan list in its corresponding hab device is empty.

However, when a HAB open request is made, the current pchan searching
routine fails to return a NULL pchan pointer when openning a domain
with no pchan inside.

This patch corrects the routine for pchan searching to prevent any
invalid pchan pointer from being accessed.

Change-Id: I7a08fb6c22343f1fdabecd53bb27201fc59cd20f
Signed-off-by: Xiang Li <quic_lixian@quicinc.com>